requirements.txt
apache-airflow==2.8.0
airflow-provider-vineyard==0.19.2
apache-airflow-providers-mongo==3.5.0
pandas==2.0.3